refactor: import Track type from state store
