+++ apps/web/package.json
@@ -29,6 +29,8 @@
     "@testing-library/react": "^14.2.1",
     "@types/react": "^18.2.47",
     "@types/react-dom": "^18.2.17",
+    "@typescript-eslint/eslint-plugin": "^8.30.1",
+    "@typescript-eslint/parser": "^8.30.1",
     "@vitejs/plugin-react": "^4.4.1",
     "autoprefixer": "^10.4.21",
     "eslint": "^8.57.0",
@@ -47,8 +49,6 @@
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.17",
     "typescript": "~5.8.3",
-    "@typescript-eslint/eslint-plugin": "^8.30.1",
-    "@typescript-eslint/parser": "^8.30.1",
     "typescript-eslint": "^8.30.1",
     "vite": "^6.3.5",
     "vitest": "^3.1.3"
