+++ apps/web/src/components/timeline/TrackRow.tsx
@@ -1,22 +1,24 @@
 import * as React from 'react'
 import { Lock, Unlock, Eye, EyeOff, Volume2, VolumeX } from 'lucide-react'
-import { type Track, useTimelineStore } from '../../stores/timelineStore'
+import { useTimelineStore, type Track } from '../../state/timelineStore'
 
 interface TrackRowProps {
   track: Track
 }
 
 export const TrackRow: React.FC<TrackRowProps> = ({ track }) => {
-  const updateTrack = useTimelineStore((s) => s.updateTrack)
+  // `updateTrack` is not implemented in the current timeline store. Stub no-op
+  // handlers so the legacy component continues to render without errors.
+  const updateTrack = React.useCallback(() => {}, [])
   const isAudio = track.type === 'audio'
 
   const toggleLock = React.useCallback(() => {
-    updateTrack(track.id, { locked: !track.locked })
-  }, [updateTrack, track])
+    updateTrack()
+  }, [updateTrack])
 
   const toggleMute = React.useCallback(() => {
-    updateTrack(track.id, { muted: !track.muted })
-  }, [updateTrack, track])
+    updateTrack()
+  }, [updateTrack])
 
   const LockIcon = track.locked ? Lock : Unlock
   const MuteIcon = track.muted
