+++ apps/web/src/tests/stores/timelineStoreBasic.test.ts
@@ -1,10 +1,20 @@
 import { describe, it, beforeEach, expect } from 'vitest'
 import { act } from '@testing-library/react'
 
-import { useTimelineStore } from '../../stores/timelineStore'
+import { useTimelineStore } from '../../state/timelineStore'
 
 const resetStore = () => {
-  useTimelineStore.setState({ tracks: [], clips: {}, currentTime: 0, zoom: 100 })
+  useTimelineStore.setState({
+    clipsById: {},
+    tracks: [],
+    durationSec: 0,
+    currentTime: 0,
+    followPlayhead: true,
+    inPoint: null,
+    outPoint: null,
+    beats: [],
+    selectedClipIds: [],
+  })
 }
 
 describe('timeline store basic actions', () => {
@@ -14,31 +24,30 @@ describe('timeline store basic actions', () => {
     let id = ''
     act(() => {
       id = useTimelineStore.getState().addClip({
-        trackId: 't1',
         start: 0,
-        duration: 2,
-        mediaId: 'm1',
+        end: 2,
+        lane: 0,
       })
     })
-    const clip = useTimelineStore.getState().clips[id]
-    expect(clip).toMatchObject({ trackId: 't1', start: 0, duration: 2, mediaId: 'm1' })
+    const clip = useTimelineStore.getState().clipsById[id]
+    expect(clip).toMatchObject({ start: 0, end: 2, lane: 0 })
   })
 
   it('updateClip modifies existing clip', () => {
     let id = ''
     act(() => {
-      id = useTimelineStore.getState().addClip({ trackId: 't1', start: 0, duration: 2, mediaId: 'm1' })
+      id = useTimelineStore.getState().addClip({ start: 0, end: 2, lane: 0 })
     })
     act(() => {
       useTimelineStore.getState().updateClip(id, { start: 1 })
     })
-    expect(useTimelineStore.getState().clips[id].start).toBe(1)
+    expect(useTimelineStore.getState().clipsById[id].start).toBe(1)
   })
 
-  it('setCurrentTime clamps to zero', () => {
+  it('setCurrentTime updates value', () => {
     act(() => {
       useTimelineStore.getState().setCurrentTime(-5)
     })
-    expect(useTimelineStore.getState().currentTime).toBe(0)
+    expect(useTimelineStore.getState().currentTime).toBe(-5)
   })
 })
